Add danio model hints
