+++ docs/index.md
@@ -24,6 +24,7 @@ Danio is a ORM for python asyncio world.It is designed to make getting easy and 
 * complex operation like bulk create, upsert, create or update and so on
 * assist model schema migration
 * support MySQL/PostgreSQL/SQLite
+* hints generation
 
 ## install
 
@@ -46,33 +47,47 @@ db = danio.Database(
 
 @dataclasses.dataclass
 class User(danio.Model):
+    # auto generated by danio:
+    # --------------------Danio Hints--------------------
+    # TABLE NAME: user
+    # TABLE IS MIGRATED!
+    ID: typing.ClassVar[danio.Field]  # "id" serial PRIMARY KEY NOT NULL
+    NAME: typing.ClassVar[danio.Field]  # "name" varchar(255)  NOT NULL
+    AGE: typing.ClassVar[danio.Field]  # "age" int  NOT NULL
+    CREATED_AT: typing.ClassVar[
+        danio.Field
+    ]  # "created_at" timestamp without time zone  NOT NULL
+    UPDATED_AT: typing.ClassVar[
+        danio.Field
+    ]  # "updated_at" timestamp without time zone  NOT NULL
+    GENDER: typing.ClassVar[danio.Field]  # "gender" int  NOT NULL
+    # --------------------Danio Hints--------------------
+
     class Gender(enum.Enum):
         MALE = 0
         FEMALE = 1
         OTHER = 2
 
-    name: str = danio.field(
-        danio.CharField,
-        comment="User name",
-        default=danio.CharField.NoDefault,
-    )
-    age: int = danio.field(danio.IntField)
-    created_at: datetime.datetime = danio.field(
-        danio.DateTimeField,
-        comment="when created",
-    )
-    updated_at: datetime.datetime = danio.field(
-        danio.DateTimeField,
-        comment="when created",
-    )
-    gender: Gender = danio.field(danio.IntField, enum=Gender, default=Gender.FEMALE)
+    id: typing.Annotated[int, danio.IntField(primary=True, type="serial")] = 0
+    name: typing.Annotated[str, danio.CharField(comment="User name")] = ""
+    age: typing.Annotated[int, danio.IntField] = 0
+    created_at: typing.Annotated[
+        datetime.datetime,
+        danio.DateTimeField(type="timestamp without time zone", comment="when created"),
+    ] = dataclasses.field(default_factory=datetime.datetime.now)
+    updated_at: typing.Annotated[
+        datetime.datetime,
+        danio.DateTimeField(type="timestamp without time zone", comment="when updated"),
+    ] = dataclasses.field(default_factory=datetime.datetime.now)
+    gender: typing.Annotated[Gender, danio.IntField(enum=Gender)] = Gender.MALE
 
-    async def before_create(self, **kwargs):
-        # user_count += 1
-        await super().before_create(**kwargs)
 
-    async def before_update(self, **kwargs):
+    async def before_create(self, validate=True):
+        await super().before_create(validate=True)
+
+    async def before_update(self, validate=True):
         self.updated_at = datetime.datetime.now()
+        await super().before_update(validate=True)
 
     async def validate(self):
         await super().validate()
@@ -87,22 +102,22 @@ class User(danio.Model):
 
 # base CRUD
 user = await User(name="batman").save()
-user = await User.where(User.name == "batman").fetch_one()
+user = await User.where(User.NAME == "batman").fetch_one()
 user.gender = User.Gender.MALE
 await user.save()
 await user.delete()
 # sql chain
-await User.where(User.name != "").limit(10).fetch_all()
+await User.where(User.NAME != "").limit(10).fetch_all()
 # multi where condition
-await User.where(User.id != 1, User.name != "").fetch_all()
-await User.where(User.id != 1).where(User.name != "").fetch_all()
-await User.where(User.id <= 10, User.id >= 20, is_and=False).fetch_all()
+await User.where(User.ID != 1, User.NAME != "").fetch_all()
+await User.where(User.ID != 1).where(User.NAME != "").fetch_all()
+await User.where(User.ID <= 10, User.ID >= 20, is_and=False).fetch_all()
 # complicated expression
-await User.where(User.id == 1).update(age=(User.age + 1) / (User.age / 12) - 2)
-await User.where((User.age + 1) == 3).fetch_all()
+await User.where(User.ID == 1).update(age=(User.AGE + 1) / (User.AGE / 12) - 2)
+await User.where((User.AGE + 1) == 3).fetch_all()
 # complicated sql operation
-await User.where(User.id == u.id).update(
-    age=User.age.case(User.age > 10, 1, default=18).case(User.age <= 0, 10)
+await User.where(User.ID == u.id).update(
+    age=User.AGE.case(User.AGE > 10, 1, default=18).case(User.AGE <= 0, 10)
 )
 created, updated = await UserProfile.upsert(
     [
@@ -116,9 +131,9 @@ await User.bulk_update(await User.fetch_all())
 await User.bulk_delete(await User.fetch_all())
 # shortcut
 user, created = await User(id=1, name="created?").get_or_create(
-    key_fields=(User.id,)
+    key_fields=(User.ID,)
 )
 user, created, updated = await User(id=2, name="updated?").create_or_update(
-    key_fields=(User.id,)
+    key_fields=(User.ID,)
 )
 ```
